kopuz: split app startup helpers
